requirements.txt
# Python dependencies
blinker==1.9.0
click==8.3.0
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymysql==1.1.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
